+++ CHANGES.md
@@ -1,1 +1,7 @@
 # Changelog
+
+## implement multilingual greet function
+- Implemented at 2026-06-25T17:56:36Z
+
+## add error handling
+- Implemented at 2026-06-25T17:57:41Z